+++ code-types/TypeConverter.js
@@ -101,6 +101,7 @@ var TypeConverter;
             case "decimal":
             case "real":
             case "number":
+            case "integer":
                 return true;
             default:
                 return false;